flask/requirements.txt
av==10.0.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.25.2
coloredlogs==15.0.1
ctranslate2==3.13.0
docopt==0.6.2
faster-whisper==0.5.1
ffmpeg-python==0.2.0
filelock==3.9.0
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
flatbuffers==23.5.9
fsspec==2023.5.0
future==0.18.3
greenlet==2.0.2
gunicorn==20.1.0
huggingface-hub==0.14.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
lit==15.0.7
llvmlite==0.40.1
MarkupSafe==2.1.2
more-itertools==9.1.0
mosestokenizer==1.2.1
mpmath==1.3.0
nltk==3.8.1
numpy==1.24.2
onnxruntime==1.14.1
openfile==0.0.7
packaging==23.0
pika==1.3.2
protobuf==4.23.1
psycopg2-binary==2.9.5
pyetherpadlite==1.1.1
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
# scipy==1.10.1
six==1.16.0
soupsieve==2.4
SQLAlchemy==2.0.7
sympy==1.12
tokenizers==0.13.2
toolwrapper==2.1.0
tqdm==4.65.0
transformers==4.26.1
typing_extensions==4.5.0
uctools==1.3.0
urllib3==1.26.14
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
